requirements.txt
pandas
numpy
polars
black
torch
ipykernel
ipywidgets
pathlib
seaborn
dill
optuna
deap
pandas-datareader
scikit-learn
mlflow
pyyaml